@@ -1,26 +1,21 @@
 name: postgres_driver
 description: A starting point for Dart libraries or applications.
-# version: 1.0.0
-# homepage: https://www.example.com
-# author: kernel <email@example.com>
 
 environment:
-  sdk: '>=2.10.4 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  ffi: 1.0.0
-  logging: 0.11.3+2
-  resource_pool:
-    git: https://github.com/radzish/resource_pool.git
+  ffi: ^1.1.2
+  logging: ^1.0.2
   db_context_lib:
     git:
       url: https://github.com/radzish/db_context.git
       path: db_context_lib
 
 dev_dependencies:
-  pedantic: ^1.8.0+1
-  test: ^1.9.4
-  process_run: 0.10.6+1
+  pedantic: ^1.11.1
+  test: ^1.18.2
+  process_run: ^0.12.2+1
   db_context_gen:
     git:
       url: https://github.com/radzish/db_context.git
@@ -28,6 +23,12 @@ dev_dependencies:
 
 dependency_overrides:
   db_context_lib:
-    path: ../../db_context/db_context_lib
+    git:
+      url: https://github.com/radzish/db_context.git
+      path: db_context_lib
+      ref: b56418f0
   db_context_gen:
-    path: ../../db_context/db_context_gen
+    git:
+      url: https://github.com/radzish/db_context.git
+      path: db_context_gen
+      ref: b56418f0
